Mount the WasteCarriersEngine and start configuration
We need to move a lot of stuff over from the renewals app, so we'll start with some initializers.

Current status is the app looks right and mounts the engine, but can't connect to databases. This is a safety commit before rebuilding Vagrant box to check if this is the cause.