Add branch to WCR renewals Gemfile entry
Now that the project is an engine and which we reference in the front-office and back-office apps when developing locally we often need to make changes to this project, which we want to see in the apps when run.

To support this we can use a feature in bundler to tell it that a gem is actually stored locally, rather than always pulling from the original source; `bundle config --local local.my_gem`.

However this option is only supported if a branch has also been defined in the entry. We also believe its useful to specify the branch in the gemfile, so other developers can be clear on what is being referenced in the source repo.

N.B. For more details on local gem config see

https://cruikshanks.co.uk/blog/local-gems-for-local-projects/
https://rossta.net/blog/how-to-specify-local-ruby-gems-in-your-gemfile.html