Enable force_ssl in production
We found when running the app in production mode it assumed SSL was being used and would therefore force a redirect to HTTPS. Setting this value to false     would not elicit this behaviour.

However it appears we need to ensure `force_ssl` is enabled to ensure that our cookies are marked as secure. So to make it explicit we are uncommenting this  line and ensuring `force_ssl` is enabled when the app is run in production.