Add test & update code for handover from frontend
A user will come to the front-office (as it currently is) from the frontend, either having entered the registration number they wish to renew, or having signed into the frontend and selected to renew a registration.

This test covers ensuring that handover works i.e. they will come to /fo/renew/CBDU12345, and if not already logged in will be redirected to sign_in, which when complete should forward them onto /fo/renew/CBDU12345.