Handle unauth. user going to /fo
Prior to this change an unauthorised user hitting `/fo` would just be redirected to the new registration start page in the frontend.

There is no issue with that, but our belief is as they are interacting with the front-office app, they should actually be redirected to the user sign in page.

To support this behaviour we have created a dashboards controller, as the behaviour and setup is very similar to the back office in this regard.

The difference is that at this time the index action on the controller (once authenticated) will redirect the user to frontend user dashboard, whereas in the future once we have implemented a new dashboard it would simply render it.