Test converting active upper tier regs to transient objects
https://eaflood.atlassian.net/browse/WC-145

As part of the design of the new renewals service, when a user comes to renew their registration we take the existing record and generate a transient_registration (a new record) from it.

This is so they can update their details during the renewal process without affecting the existing registration, and retain their existing reg number. Once the process is complete, the transient object switches to being their renewed registration.

However, due to the less structured nature of the database, we cannot assume that all existing upper tier registrations have all the fields, properties and sub documents we might expect. Therefore we need a mechanism to test the conversion from a registration to a transient_registration to ensure it can handle existing registrations.