Create rake task to convert matching registrations to transient_regs
This task finds all registrations where the tier is UPPER and the metaData.status is ACTIVE and attempts to use them to create a transient_registration and save it to the database.

Details of any errors are saved to a new log file.

Worth noting this does make a mess of the DB.