Fix reg to trans test to override Mongo timeout
https://eaflood.atlassian.net/browse/WC-145

When running this task against actual production data we found that it keeps erroring, but at different points. The error is always

```
rake aborted!
Mongo::Error::OperationFailure: cursor id 31540126646 not found (43)
/vagrant/waste-carriers-front-office/lib/tasks/create_transient_registrations.rake:16:in `block in <top (required)>'
/home/vagrant/.rbenv/versions/2.4.2/bin/bundle:23:in `load'
/home/vagrant/.rbenv/versions/2.4.2/bin/bundle:23:in `<main>'
Tasks: TOP => create_transient_registrations
(See full trace by running task with --trace)
```

Initial Googling seemed to indicate this was caused by a default timeout MongoDb has on cursors of 10  minutes (simply iterating through it does not seem to count as activity). So we trialled changing the query to include `no_cursor_timeout` and this allowed many more to complete, but we still eventually errored.

Next we incresed the number of CPU's being used in the VM this task is running in, and again got further before finally erroring.

So in the end we re-wrote the task to use paging to iterate through those registrations which are renewable and attempt to convert them. That combined with `no_cursor_timeout` and more CPU's meant we finally completed a run where we test converted all 122K eligible registrations.

🔥 WIP 🔥