Update CSP to include tagmanager.google.com
https://eaflood.atlassian.net/browse/WC-476

Our integration with Google analytics is working, but we faced errors when trying to enable getting stats on what errors get raised within the service.

The issue was tracked down to the fact our current Content Security policy does not allow scripts from tagmanager.google.com to be run.

This updates the CSP to include it.

🔥 WIP 🔥