Add config for renewal grace window
https://eaflood.atlassian.net/browse/WC-512

We have added to [Waste-carrier-renewals](https://github.com/DEFRA/waste-carriers-renewals) a 'grace window' which allows users of registrations that have expired to still renew, if the current date is within the window.

However to support this new feature we have to add config to the app which hosts the engine hence this change.

🎉 WIP 🎉