After user confirms password, redirect to frontend
Once the user has confirmed their password, we don't ask them to sign in again. Instead we redirect them to the frontend. In the development environment, this displays a test page - however in production it should confirm they are logged in and redirect them to the dashboard, where they can see their registration.