Set up registration dashboard
https://eaflood.atlassian.net/browse/RUBY-14

This feature comes out of our attempt to fix time-outs which occur when a user with a large number of registrations tries to view their dashboard. The frontend requests allll their registrations from the Java service. This takes a long time and means the request times out, so the user just gets an error.

We attempted to fix this by introducing pagination: https://github.com/DEFRA/waste-carriers-frontend/pull/178 However, this didn't resolve the underlying issue of only getting registration data from the Java service instead of directly from the database, so the error still occurred.

Rather than re-engineer the frontend and the Java service to make this work, it seemed more expedient to just re-implement the dashboard in the front-office, this time loading data directly from the database, and including pagination to avoid timeouts.