Add logic to display renew link
The frontend handles this with a separate service which does various checks on the registration.

Most of these checks (and more) exist in the engine - however, they're all currently on the transient_registration. So rather than duplicate that code, we test an existing transient_registration, or instantiate (but don't persist) a new one just for testing.

I think when we tackle https://github.com/DEFRA/waste-carriers-renewals/issues/324 this should probably be refactored.