Add support for mocking external services
We have created a new engine in the DefraRuby namespace called [defra_ruby_mocks](https://github.com/DEFRA/defra-ruby-mocks).

Going forward this will allow us to mock external services like Companies House, and in the future Worldpay.

We only really need to do this when running performance tests. We want to put as much of the service under load as possible, but at the same time we don't want to get into trouble for overloading these external services, or worst case get accused of a DDOS attack.

So by providing an alternate end point that can accept the same request, and return what the service needs, we can replace them when we need to.

For reference we only need to bring mocking support into one app, because we only need one version of the mock endpoint running. And we chose the front-office because we shouldn't get any issues with access permissions from things like white listing.

So this change adds the mocks engine to the app, including the initializer, which for now will give us the ability to mock companies house.