Add config for the mock delay
We actually have an existing environment variable we use for setting the delay to use with the existing mocking functionality.

Should we come to want to continue to use it, the initializer will need to look for and pull values from the env var, as just use the default built into the mock engine.