Mount mocks in the /fo path
Working locally just mounting the mocks at `/mocks` works fine. This is because locally we can distinguish between the frontend and the front-office using ports (3000 vs 3002).

However in our AWS environments this is not the case. Both share the same domain of https://waste-carriers-dev.aws.defra.cloud. So for the environment (Nginx) to understand which app to route traffic through to it relies on the use of the `/fo` path.

If the request is for domain + `/fo` route it to the front-office app. If it's just for the domain, route it to the frontend app.

So for the frontend and the back-office to be able to make requests to the mocks loaded into the front-office, the mocks must be mounted in the `/fo` path.