Remove delete link from user dashboard
https://eaflood.atlassian.net/browse/RUBY-807

We have built a new 'Cease or Revoke' feature into the back-office to replace the existing 'De-register' and 'Revoke' functions in the old backend.

As part of the rebuild it was agreed with NCCC that we would not build a version for external users; this would be for internal users only in the future. In fact, it was requested this ability be removed from the external facing sites because of the issues it causes when users mistakenly 'Delete' their registration.

We forgot to remove the old external 'Delete' link at the time. This change rectifies that oversight.