Add certificate controller and route
This basically mimics exactly what we have in the back office.

This does mean some duplication that could have been avoided if we moved this logic to the engine instead - however since the implementation of this feature is relatively temporary, I think this makes it easier to remove down the line.