Update dashboard 'change password' link
https://eaflood.atlassian.net/browse/RUBY-1129

This previously linked to the password change functionality in the old frontend. It should go to the front office instead.

It turns out that the existing "reset password" function for users who are not logged in could not be reused here - this is specifically for un-authed users, and if you want to use Devise to let logged-in users edit their details, it comes as part of the 'registerable' module. We don't want all of the baggage that comes with that (ie users being able to create their own accounts at will, or edit other account values) so after consulting the docs, the most straightforward option was just to make our own.

Reference: https://github.com/heartcombo/devise/wiki/How-To:-Allow-users-to-edit-their-password