Redirect to /fo after password reset
https://eaflood.atlassian.net/browse/RUBY-1129

When we implemented the password edit in [PR #559](https://github.com/DEFRA/waste-carriers-front-office/pull/559) we had it redirect to root after the POST.

Locally this would redirect to the dashboard. But when testing in AWS it was redirecting to `/fo/start`. When investigating this was because `root_path()` was returning `/` and in our AWS environments if `/fo` is not at the start of the path it will direct the request to the old frontend app.

However, that now also has an automatic redirect to `/fo/start`. So this is why it looks like the change password screen is redirecting to the new registration start screen.

This change tells the passwords controller to redirect to `/fo` after submission. If the user is signed in this will return them to the dashboard. If not it will send them to the sign in page.